Cleanup and consolidate supporting GraphQL code
