--- github/util_v4_consts.go
@@ -1,18 +1,0 @@
-package github
-
-const (
-	PROTECTION_DISMISSES_STALE_REVIEWS         = "dismiss_stale_reviews"
-	PROTECTION_IS_ADMIN_ENFORCED               = "enforce_admins"
-	PROTECTION_PATTERN                         = "pattern"
-	PROTECTION_REQUIRED_APPROVING_REVIEW_COUNT = "required_approving_review_count"
-	PROTECTION_REQUIRED_STATUS_CHECK_CONTEXTS  = "contexts"
-	PROTECTION_REQUIRES_APPROVING_REVIEWS      = "required_pull_request_reviews"
-	PROTECTION_REQUIRES_CODE_OWNER_REVIEWS     = "require_code_owner_reviews"
-	PROTECTION_REQUIRES_COMMIT_SIGNATURES      = "require_signed_commits"
-	PROTECTION_REQUIRES_STATUS_CHECKS          = "required_status_checks"
-	PROTECTION_REQUIRES_STRICT_STATUS_CHECKS   = "strict"
-	PROTECTION_RESTRICTS_PUSHES                = "push_restrictions"
-	PROTECTION_RESTRICTS_REVIEW_DISMISSALS     = "dismissal_restrictions"
-
-	REPOSITORY_ID = "repository_id"
-)
